requirements.txt
praw==7.7.1
selenium==4.15.2
python-dotenv==1.0.0
playwright==1.40.0
gtts==2.4.0
moviepy==1.0.3
Pillow==10.1.0
numpy==1.26.2